+++ R/request_matcher_registry.R
@@ -1,107 +1,55 @@
-#' @title RequestMatcherRegistry
-#' @description handles request matchers
-#' @export
-RequestMatcherRegistry <- R6::R6Class(
-  'RequestMatcherRegistry',
-  public = list(
-    #' @field registry initialze registry list with a request, or leave empty
-    registry = NULL,
-    #' @field default_matchers request matchers to use. default: method, uri
-    default_matchers = NULL,
+request_matches <- function(req1, req2, match_requests_on) {
+  for (matcher in match_requests_on) {
+    match <- request_matches_one(matcher, req1, req2)
+    vcr_log_sprintf(
+      "    %s %s: current request [%s] vs [%s]",
+      matcher,
+      if (match) "matched" else "did not match",
+      request_summary(req1, match_requests_on),
+      request_summary(req2, match_requests_on)
+    )
 
-    #' @description Create a new RequestMatcherRegistry object
-    #' @param registry initialze registry list with a request, or leave empty
-    #' @param default_matchers request matchers to use. default: method, uri
-    #' @return A new `RequestMatcherRegistry` object
-    initialize = function(
-      registry = list(),
-      default_matchers = list('method', 'uri')
-    ) {
-      self$registry <- registry
-      self$default_matchers <- default_matchers
-      self$register_built_ins()
-    },
-
-    #' @description Register a custom matcher
-    #' @param name matcher name
-    #' @param func function that describes a matcher, should return
-    #' a single boolean
-    #' @return no return; registers the matcher
-    register = function(name, func) {
-      if (name %in% self$registry) {
-        warning(
-          sprintf(
-            "There is already a vcr request matcher registered for %s. Overriding it.",
-            name
-          )
-        )
-      }
-      self$registry[[name]] <- Matcher$new(func = func)
-    },
-
-    #' @description Register all built in matchers
-    #' @return no return; registers all built in matchers
-    #' @note r1=from new request; r2=from recorded interaction
-    register_built_ins = function() {
-      self$register("method", function(r1, r2) r1$method == r2$method)
-      self$register(
-        "uri",
-        function(r1, r2)
-          identical(
-            curl::curl_unescape(encode_uri(r1$uri, flip = TRUE)),
-            curl::curl_unescape(r2$uri)
-          )
-      )
-      self$register(
-        "body",
-        function(r1, r2) identical(as.character(r1$body), as.character(r2$body))
-      )
-      self$register(
-        'headers',
-        function(r1, r2) identical(r1$headers, r2$headers)
-      )
-      self$register("host", function(r1, r2) identical(r1$host, r2$host))
-      self$register(
-        "path",
-        function(r1, r2)
-          identical(sub("/$", "", r1$path), sub("/$", "", r2$path))
-      )
-      self$register(
-        "query",
-        function(r1, r2)
-          identical(
-            curl::curl_unescape(r1$query),
-            curl::curl_unescape(r2$query)
-          )
-      )
-      self$try_to_register_body_as_json()
-    },
-
-    #' @description Try to register body as JSON
-    #' @param r1,r2 [Request] class objects
-    #' @return no return; registers the matcher
-    try_to_register_body_as_json = function(r1, r2) {
-      if (!requireNamespace("jsonlite", quietly = TRUE)) {
-        stop("please install jsonlite", call. = FALSE)
-      }
-
-      self$register("body_as_json", function(r1, r2) {
-        tc <- tryCatch(
-          identical(jsonlite::fromJSON(r1$body), jsonlite::fromJSON(r2$body)),
-          error = function(e) e
-        )
-        if (inherits(tc, "error")) FALSE else tc
-      })
+    if (!match) {
+      return(FALSE)
     }
-  )
-)
+  }
 
-# generate matcher functions
-Matcher <- R6::R6Class(
-  'Matcher',
-  public = list(
-    func = NULL,
-    initialize = function(func) self$func <- func,
-    matches = function(...) self$func(...)
+  TRUE
+}
+
+request_matches_one <- function(type, req1, req2) {
+  switch(
+    type,
+    method = req1$method == req2$method,
+    uri = identical(
+      url_without_port(encode_uri(req1$uri, flip = TRUE)),
+      url_without_port(req2$uri)
+    ),
+    uri_with_port = identical(
+      curl::curl_unescape(encode_uri(req1$uri, flip = TRUE)),
+      curl::curl_unescape(req2$uri)
+    ),
+    body = identical(req1$body, req2$body),
+    headers = identical(req1$headers, req2$headers),
+    host = identical(url_host(req1$uri), url_host(req2$uri)),
+    path = identical(url_path(req1$uri), url_path(req2$uri)),
+    query = identical(url_query(req1$uri), url_query(req2$uri)),
+    cli::cli_abort("Unsupported request matcher {.str {type}}.")
   )
-)
+}
+
+url_path <- function(x) {
+  sub("/$", "", curl::curl_parse_url(x)$path)
+}
+url_query <- function(x) {
+  curl::curl_parse_url(x)$query
+}
+url_host <- function(x) {
+  curl::curl_parse_url(x)$host
+}
+url_without_port <- function(x) {
+  url <- curl::curl_parse_url(x)
+  url$port <- NULL
+  url$url <- NULL
+  url
+}