+++ R/utils.R
@@ -81,29 +81,15 @@ dir_create <- function(path) {
   dir.create(path, showWarnings = FALSE, recursive = TRUE)
 }
 
-to_base64 <- function(x) {
-  x <- jsonlite::base64_enc(x)
-
-  # Split into lines of 80 characters
-  chunk_size <- 80
-  length <- nchar(x)
-  if (length < chunk_size) {
-    return(x)
-  }
-
-  num_chunks <- ceiling(length / 80)
-  start_positions <- seq(1, by = 80, length.out = num_chunks)
-  end_positions <- pmin(start_positions + chunk_size - 1, length)
-
-  lines <- substring(x, start_positions, end_positions)
-  paste0(lines, collapse = "\n")
-}
-
-from_base64 <- function(x) {
-  x <- gsub("[\r\n]", "", x)
-  jsonlite::base64_dec(x)
-}
-
 cur_time <- function(tz = "") {
   format(Sys.time(), format = "%Y-%m-%d %H:%M:%S", tz = tz)
+}
+
+
+pkg_versions <- function() {
+  paste(
+    paste0("vcr/", utils::packageVersion("vcr")),
+    paste0("webmockr/", utils::packageVersion("webmockr")),
+    sep = ", "
+  )
 }